requirements.txt
apache-beam[gcp]==2.28.0
frozendict==1.2
pandas~=1.0.0
scikit-learn~=0.22.1
tensorflow~=2.2.0
tf-models-official~=2.2.0
kaggle<1.5.8
tensorflow-probability==0.10.1
tf_slim~=1.1.0
nltk~=3.5
beautifulsoup4==4.9.3
html5lib==1.1
gensim~=3.8.3
lxml~=4.6.0
tensorboard=2.2.2
hydra-core~=1.3.2
omegaconf~=2.3.0
protobuf~=3.20.3